+++ backend/src/main/java/com/golden_dobakhe/HakPle/domain/post/comment/comment/entity/Comment.java
@@ -3,7 +3,7 @@ package com.golden_dobakhe.HakPle.domain.post.comment.comment.entity;
 import com.golden_dobakhe.HakPle.domain.post.post.entity.Board;
 import com.golden_dobakhe.HakPle.domain.user.user.entity.User;
 import com.golden_dobakhe.HakPle.global.entity.BaseEntity;
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
