+++ backend/src/main/java/com/golden_dobakhe/HakPle/domain/post/post/service/impl/BoardServiceImpl.java
@@ -10,7 +10,7 @@ import com.golden_dobakhe.HakPle.domain.post.post.service.BoardService;
 
 import com.golden_dobakhe.HakPle.domain.user.user.entity.User;
 import com.golden_dobakhe.HakPle.domain.user.user.repository.UserRepository;
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
