+++ backend/src/main/java/com/golden_dobakhe/HakPle/domain/user/user/entity/User.java
@@ -2,7 +2,7 @@ package com.golden_dobakhe.HakPle.domain.user.user.entity;
 
 import com.golden_dobakhe.HakPle.domain.resource.image.entity.Image;
 import com.golden_dobakhe.HakPle.global.entity.BaseEntity;
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
