+++ backend/src/main/java/com/golden_dobakhe/HakPle/domain/user/user/service/UserService.java
@@ -6,7 +6,7 @@ import com.golden_dobakhe.HakPle.domain.user.user.dto.UserDTO;
 import com.golden_dobakhe.HakPle.domain.user.user.entity.User;
 import com.golden_dobakhe.HakPle.domain.user.user.exception.UserException;
 import com.golden_dobakhe.HakPle.domain.user.user.repository.UserRepository;
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -24,12 +24,12 @@ public class UserService {
     public void register(UserDTO userDTO) {
         // 사용자 이름 중복 확인
         if (userRepository.existsByUserName(userDTO.getUserName())) {
-            throw new IllegalArgumentException("아이디가 이미 사용 중입니다.");
+            throw new UserException(UserErrorCode.USERNAME_DUPLICATE);
         }
 
         // 닉네임 중복 확인
         if (userRepository.existsByNickName(userDTO.getNickName())) {
-            throw new IllegalArgumentException("닉네임이 이미 사용 중입니다.");
+            throw new UserException(UserErrorCode.NICKNAME_DUPLICATE);
         }
 
         // User 엔티티로 변환 및 저장
