+++ backend/src/main/java/com/golden_dobakhe/HakPle/security/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.golden_dobakhe.HakPle.security;
 
 import com.golden_dobakhe.HakPle.domain.user.user.entity.User;
 import com.golden_dobakhe.HakPle.domain.user.user.repository.UserRepository;
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
