+++ backend/src/main/java/com/golden_dobakhe/HakPle/security/jwt/JwtAuthFilter.java
@@ -3,7 +3,7 @@ package com.golden_dobakhe.HakPle.security.jwt;
 //시큐리티에게 jwt를 넘겨주기 위한 필터
 
 
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import com.golden_dobakhe.HakPle.security.CustomUserDetailsService;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
