+++ backend/src/main/java/com/golden_dobakhe/HakPle/security/jwt/JwtTokenizer.java
@@ -1,6 +1,6 @@
 package com.golden_dobakhe.HakPle.security.jwt;
 
-import com.golden_dobakhe.HakPle.global.entity.Status;
+import com.golden_dobakhe.HakPle.global.Status;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Component;
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
 import java.util.Date;
-import java.util.List;
 
 @Component
 public class JwtTokenizer {
